requirements.txt
pyspark
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-storage
pandas
pyarrow
fastavro
confluent-kafka
attrs
httpx
cachetools
authlib